+++ libcaf_core/caf/detail/behavior_impl.hpp
@@ -31,7 +31,6 @@
 #include "caf/detail/type_traits.hpp"
 #include "caf/intrusive_ptr.hpp"
 #include "caf/make_counted.hpp"
-#include "caf/match_result.hpp"
 #include "caf/message.hpp"
 #include "caf/none.hpp"
 #include "caf/optional.hpp"
@@ -62,10 +61,9 @@ public:
 
   explicit behavior_impl(timespan tout);
 
-  match_result invoke_empty(detail::invoke_result_visitor& f);
+  bool invoke_empty(detail::invoke_result_visitor& f);
 
-  virtual match_result invoke(detail::invoke_result_visitor& f, message& xs)
-    = 0;
+  virtual bool invoke(detail::invoke_result_visitor& f, message& xs) = 0;
 
   optional<message> invoke(message&);
 
@@ -123,15 +121,13 @@ public:
     // nop
   }
 
-  virtual match_result invoke(detail::invoke_result_visitor& f,
-                              message& xs) override {
+  virtual bool invoke(detail::invoke_result_visitor& f, message& xs) override {
     return invoke_impl(f, xs, std::make_index_sequence<sizeof...(Ts)>{});
   }
 
   template <size_t... Is>
-  match_result invoke_impl(detail::invoke_result_visitor& f, message& msg,
-                           std::index_sequence<Is...>) {
-    auto result = match_result::no_match;
+  bool invoke_impl(detail::invoke_result_visitor& f, message& msg,
+                   std::index_sequence<Is...>) {
     auto dispatch = [&](auto& fun) {
       using fun_type = std::decay_t<decltype(fun)>;
       using trait = get_callable_trait_t<fun_type>;
@@ -141,18 +137,16 @@ public:
         using fun_result = decltype(detail::apply_args(fun, xs));
         if constexpr (std::is_same<void, fun_result>::value) {
           detail::apply_args(fun, xs);
-          result = f.visit(unit) ? match_result::match : match_result::skip;
+          f(unit);
         } else {
           auto invoke_res = detail::apply_args(fun, xs);
-          result = f.visit(invoke_res) ? match_result::match
-                                       : match_result::skip;
+          f(invoke_res);
         }
         return true;
       }
       return false;
     };
-    static_cast<void>((dispatch(std::get<Is>(cases_)) || ...));
-    return result;
+    return (dispatch(std::get<Is>(cases_)) || ...);
   }
 
   void handle_timeout() override {