+++ libcaf_core/src/detail/behavior_impl.cpp
@@ -28,9 +28,8 @@ namespace {
 
 class combinator final : public behavior_impl {
 public:
-  match_result invoke(detail::invoke_result_visitor& f, message& xs) override {
-    auto x = first->invoke(f, xs);
-    return x == match_result::no_match ? second->invoke(f, xs) : x;
+  bool invoke(detail::invoke_result_visitor& f, message& xs) override {
+    return first->invoke(f, xs) || second->invoke(f, xs);
   }
 
   void handle_timeout() override {
@@ -53,20 +52,12 @@ class maybe_message_visitor : public detail::invoke_result_visitor {
 public:
   optional<message> value;
 
-  void operator()() override {
-    value = message{};
-  }
-
   void operator()(error& x) override {
     value = make_message(std::move(x));
   }
 
   void operator()(message& x) override {
     value = std::move(x);
-  }
-
-  void operator()(const none_t&) override {
-    (*this)();
   }
 };
 
@@ -84,27 +75,20 @@ behavior_impl::behavior_impl(timespan tout) : timeout_(tout) {
   // nop
 }
 
-match_result behavior_impl::invoke_empty(detail::invoke_result_visitor& f) {
+bool behavior_impl::invoke_empty(detail::invoke_result_visitor& f) {
   message xs;
   return invoke(f, xs);
 }
 
 optional<message> behavior_impl::invoke(message& xs) {
   maybe_message_visitor f;
-  // the following const-cast is safe, because invoke() is aware of
-  // copy-on-write and does not modify x if it's shared
-  if (!xs.empty())
-    invoke(f, xs);
-  else
-    invoke_empty(f);
-  return std::move(f.value);
+  if (invoke(f, xs))
+    return std::move(f.value);
+  return none;
 }
 
-match_result behavior_impl::invoke(detail::invoke_result_visitor& f,
-                                   message& xs) {
-  if (!xs.empty())
-    return invoke(f, xs);
-  return invoke_empty(f);
+bool behavior_impl::invoke(detail::invoke_result_visitor& f, message& xs) {
+  return invoke(f, xs);
 }
 
 void behavior_impl::handle_timeout() {