+++ libcaf_core/src/scheduled_actor.cpp
@@ -36,26 +36,26 @@ namespace caf {
 
 // -- related free functions ---------------------------------------------------
 
-result<message> reflect(scheduled_actor*, message& msg) {
+skippable_result reflect(scheduled_actor*, message& msg) {
   return std::move(msg);
 }
 
-result<message> reflect_and_quit(scheduled_actor* ptr, message& msg) {
+skippable_result reflect_and_quit(scheduled_actor* ptr, message& msg) {
   error err = exit_reason::normal;
   scheduled_actor::default_error_handler(ptr, err);
   return reflect(ptr, msg);
 }
 
-result<message> print_and_drop(scheduled_actor* ptr, message& msg) {
+skippable_result print_and_drop(scheduled_actor* ptr, message& msg) {
   CAF_LOG_WARNING("unexpected message:" << msg);
   aout(ptr) << "*** unexpected message [id: " << ptr->id()
             << ", name: " << ptr->name() << "]: " << to_string(msg)
             << std::endl;
-  return sec::unexpected_message;
+  return make_error(sec::unexpected_message);
 }
 
-result<message> drop(scheduled_actor*, message&) {
-  return sec::unexpected_message;
+skippable_result drop(scheduled_actor*, message&) {
+  return make_error(sec::unexpected_message);
 }
 
 // -- implementation details ---------------------------------------------------
@@ -67,7 +67,7 @@ void silently_ignore(scheduled_actor*, T&) {
   // nop
 }
 
-result<message> drop_after_quit(scheduled_actor* self, message&) {
+skippable_result drop_after_quit(scheduled_actor* self, message&) {
   if (self->current_message_id().is_request())
     return make_error(sec::request_receiver_down);
   return make_message();
@@ -691,38 +691,26 @@ invoke_message_result scheduled_actor::consume(mailbox_element& x) {
         return invoke_message_result::consumed;
       case message_category::ordinary: {
         detail::default_invoke_result_visitor<scheduled_actor> visitor{this};
-        bool skipped = false;
         auto had_timeout = getf(has_timeout_flag);
         if (had_timeout)
           unsetf(has_timeout_flag);
-        // restore timeout at scope exit if message was skipped
-        auto timeout_guard = detail::make_scope_guard([&] {
-          if (skipped && had_timeout)
-            setf(has_timeout_flag);
-        });
-        auto call_default_handler = [&] {
-          auto f = detail::make_overload([&](auto& val) { visitor.visit(val); },
-                                         [&](skip_t&) { skipped = true; });
-          auto sres = call_handler(default_handler_, this, x.payload);
-          visit(f, sres);
-        };
-        if (bhvr_stack_.empty()) {
-          call_default_handler();
-          return !skipped ? invoke_message_result::consumed
-                          : invoke_message_result::skipped;
+        if (!bhvr_stack_.empty()) {
+          auto& bhvr = bhvr_stack_.back();
+          if (bhvr(visitor, x.content()))
+            return invoke_message_result::consumed;
         }
-        auto& bhvr = bhvr_stack_.back();
-        switch (bhvr(visitor, x.content())) {
-          default:
-            break;
-          case match_result::skip:
-            skipped = true;
-            break;
-          case match_result::no_match:
-            call_default_handler();
-        }
-        return !skipped ? invoke_message_result::consumed
-                        : invoke_message_result::skipped;
+        auto sres = call_handler(default_handler_, this, x.payload);
+        auto f = detail::make_overload(
+          [&](auto& x) {
+            visitor(x);
+            return invoke_message_result::consumed;
+          },
+          [&](skip_t&) {
+            if (had_timeout)
+              setf(has_timeout_flag);
+            return invoke_message_result::skipped;
+          });
+        return visit(f, sres);
       }
     }
     // Unreachable.
@@ -1100,19 +1088,11 @@ scheduled_actor::handle_open_stream_msg(mailbox_element& x) {
   CAF_LOG_TRACE(CAF_ARG(x));
   // Fetches a stream manager from a behavior.
   struct visitor : detail::invoke_result_visitor {
-    void operator()() override {
-      // nop
-    }
-
     void operator()(error&) override {
       // nop
     }
 
     void operator()(message&) override {
-      // nop
-    }
-
-    void operator()(const none_t&) override {
       // nop
     }
   };
@@ -1128,34 +1108,19 @@ scheduled_actor::handle_open_stream_msg(mailbox_element& x) {
     auto rp = make_response_promise();
     rp.deliver(sec::stream_init_failed);
   };
-  // Utility for invoking the default handler.
-  auto fallback = [&] {
-    auto sres = call_handler(default_handler_, this, x.payload);
-    if (holds_alternative<skip_t>(sres)) {
-      CAF_LOG_DEBUG("default handler skipped open_stream_msg:" << osm.msg);
-      return invoke_message_result::skipped;
-    } else {
-      CAF_LOG_DEBUG(
-        "default handler was called for open_stream_msg:" << osm.msg);
-      fail(sec::stream_init_failed, "dropped open_stream_msg (no match)");
-      return invoke_message_result::dropped;
-    }
-  };
   // Invoke behavior and dispatch on the result.
   auto& bs = bhvr_stack();
-  if (bs.empty())
-    return fallback();
-  auto res = (bs.back())(f, osm.msg);
-  switch (res) {
-    case match_result::no_match:
-      CAF_LOG_DEBUG("no match in behavior, fall back to default handler");
-      return fallback();
-    case match_result::match: {
-      return invoke_message_result::consumed;
-    }
-    default:
-      CAF_LOG_DEBUG("behavior skipped open_stream_msg:" << osm.msg);
-      return invoke_message_result::skipped; // nop
+  if (!bs.empty() && bs.back()(f, osm.msg))
+    return invoke_message_result::consumed;
+  CAF_LOG_DEBUG("no match in behavior, fall back to default handler");
+  auto sres = call_handler(default_handler_, this, x.payload);
+  if (holds_alternative<skip_t>(sres)) {
+    CAF_LOG_DEBUG("default handler skipped open_stream_msg:" << osm.msg);
+    return invoke_message_result::skipped;
+  } else {
+    CAF_LOG_DEBUG("default handler was called for open_stream_msg:" << osm.msg);
+    fail(sec::stream_init_failed, "dropped open_stream_msg (no match)");
+    return invoke_message_result::dropped;
   }
 }